src/requirements.txt
streamlit
pandas
openai
requests
python-dotenv

